+++ squeal-postgresql/src/Squeal/PostgreSQL/Definition.hs
@@ -98,6 +98,7 @@ import qualified GHC.Generics as GHC
 import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Logic
+import Squeal.PostgreSQL.Expression.Type
 import Squeal.PostgreSQL.List
 import Squeal.PostgreSQL.PG
 import Squeal.PostgreSQL.Query
