+++ squeal-postgresql/src/Squeal/PostgreSQL/Expression.hs
@@ -46,50 +46,7 @@ module Squeal.PostgreSQL.Expression
   , FunctionN
   , unsafeFunctionN
   , Literal (..)
-  , HasParameter (parameter)
-  , param
   , PGSubset (..)
-    -- * Types
-  , TypeExpression (..)
-  , cast
-  , PGTyped (..)
-  , typedef
-  , typetable
-  , typeview
-  , bool
-  , int2
-  , smallint
-  , int4
-  , int
-  , integer
-  , int8
-  , bigint
-  , numeric
-  , float4
-  , real
-  , float8
-  , doublePrecision
-  , money
-  , text
-  , char
-  , character
-  , varchar
-  , characterVarying
-  , bytea
-  , timestamp
-  , timestampWithTimeZone
-  , date
-  , time
-  , timeWithTimeZone
-  , interval
-  , uuid
-  , inet
-  , json
-  , jsonb
-  , vararray
-  , fixarray
-  , tsvector
-  , tsquery
     -- * Re-export
   , (&)
   , K (..)
@@ -112,12 +69,10 @@ import GHC.TypeLits
 import Numeric
 import Prelude hiding (id, (.))
 
-import qualified Data.ByteString as ByteString
 import qualified Data.Text as Text
 import qualified Data.Text.Lazy as Lazy (Text)
 import qualified Data.Text.Lazy as Lazy.Text
 import qualified GHC.Generics as GHC
-import qualified Generics.SOP as SOP
 
 import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Binary
@@ -235,46 +190,6 @@ greatest(TRUE, NULL, FALSE)
 unsafeFunctionVar :: ByteString -> FunctionVar x0 x1 y
 unsafeFunctionVar fun xs x = UnsafeExpression $ fun <> parenthesized
   (commaSeparated (renderSQL <$> xs) <> ", " <> renderSQL x)
-
-{- | A `HasParameter` constraint is used to indicate a value that is
-supplied externally to a SQL statement.
-`Squeal.PostgreSQL.PQ.manipulateParams`,
-`Squeal.PostgreSQL.PQ.queryParams` and
-`Squeal.PostgreSQL.PQ.traversePrepared` support specifying data values
-separately from the SQL command string, in which case `param`s are used to
-refer to the out-of-line data values.
--}
-class KnownNat n => HasParameter
-  (n :: Nat)
-  (params :: [NullityType])
-  (ty :: NullityType)
-  | n params -> ty where
-    -- | `parameter` takes a `Nat` using type application and a `TypeExpression`.
-    --
-    -- >>> let expr = parameter @1 int4 :: Expression outer '[] grp schemas '[ 'Null 'PGint4] from ('Null 'PGint4)
-    -- >>> printSQL expr
-    -- ($1 :: int4)
-    parameter
-      :: TypeExpression schemas ty
-      -> Expression outer commons grp schemas params from ty
-    parameter ty = UnsafeExpression $ parenthesized $
-      "$" <> renderNat @n <+> "::"
-        <+> renderSQL ty
-instance {-# OVERLAPPING #-} HasParameter 1 (ty1:tys) ty1
-instance {-# OVERLAPPABLE #-} (KnownNat n, HasParameter (n-1) params ty)
-  => HasParameter n (ty' : params) ty
-
--- | `param` takes a `Nat` using type application and for basic types,
--- infers a `TypeExpression`.
---
--- >>> let expr = param @1 :: Expression outer commons grp schemas '[ 'Null 'PGint4] from ('Null 'PGint4)
--- >>> printSQL expr
--- ($1 :: int4)
-param
-  :: forall n outer commons schemas params from grp ty
-   . (PGTyped schemas ty, HasParameter n params ty)
-  => Expression outer commons grp schemas params from ty -- ^ param
-param = parameter @n (pgtype @schemas)
 
 instance (HasUnique tab (Join outer from) row, Has col row ty)
   => IsLabel col (Expression outer commons 'Ungrouped schemas params from ty) where
@@ -457,36 +372,17 @@ instance PGSubset 'PGjsonb
 instance PGSubset 'PGtsquery
 instance PGSubset ('PGvararray ty)
 
--- When a `cast` is applied to an `Expression` of a known type, it
--- represents a run-time type conversion. The cast will succeed only if a
--- suitable type conversion operation has been defined.
---
--- | >>> printSQL $ true & cast int4
--- (TRUE :: int4)
-cast
-  :: TypeExpression schemas ty1
-  -- ^ type to cast as
-  -> Expression outer commons grp schemas params from ty0
-  -- ^ value to convert
-  -> Expression outer commons grp schemas params from ty1
-cast ty x = UnsafeExpression $ parenthesized $
-  renderSQL x <+> "::" <+> renderSQL ty
-
-{-----------------------------------------
-text
------------------------------------------}
-
 instance IsString
   (Expression outer commons grp schemas params from (null 'PGtext)) where
     fromString str = UnsafeExpression $
       "E\'" <> fromString (escape =<< str) <> "\'"
 instance IsString
   (Expression outer commons grp schemas params from (null 'PGtsvector)) where
-    fromString str = cast tsvector . UnsafeExpression $
+    fromString str = UnsafeExpression . parenthesized . (<> " :: tsvector") $
       "E\'" <> fromString (escape =<< str) <> "\'"
 instance IsString
   (Expression outer commons grp schemas params from (null 'PGtsquery)) where
-    fromString str = cast tsquery . UnsafeExpression $
+    fromString str = UnsafeExpression . parenthesized . (<> " :: tsquery") $
       "E\'" <> fromString (escape =<< str) <> "\'"
 
 instance Semigroup
@@ -511,192 +407,13 @@ instance Monoid
     mempty = fromString ""
     mappend = (<>)
 
-{-----------------------------------------
-type expressions
------------------------------------------}
-
--- | `TypeExpression`s are used in `cast`s and
--- `Squeal.PostgreSQL.Definition.createTable` commands.
-newtype TypeExpression (schemas :: SchemasType) (ty :: NullityType)
-  = UnsafeTypeExpression { renderTypeExpression :: ByteString }
-  deriving (GHC.Generic,Show,Eq,Ord,NFData)
-instance RenderSQL (TypeExpression schemas ty) where
-  renderSQL = renderTypeExpression
-
--- | The enum or composite type in a `Typedef` can be expressed by its alias.
-typedef
-  :: (Has sch schemas schema, Has td schema ('Typedef ty))
-  => QualifiedAlias sch td
-  -> TypeExpression schemas (null ty)
-typedef = UnsafeTypeExpression . renderSQL
-
--- | The composite type corresponding to a `Table` definition can be expressed
--- by its alias.
-typetable
-  :: (Has sch schemas schema, Has tab schema ('Table table))
-  => QualifiedAlias sch tab
-  -> TypeExpression schemas (null ('PGcomposite (TableToRow table)))
-typetable = UnsafeTypeExpression . renderSQL
-
--- | The composite type corresponding to a `View` definition can be expressed
--- by its alias.
-typeview
-  :: (Has sch schemas schema, Has vw schema ('View view))
-  => QualifiedAlias sch vw
-  -> TypeExpression schemas (null ('PGcomposite view))
-typeview = UnsafeTypeExpression . renderSQL
-
--- | logical Boolean (true/false)
-bool :: TypeExpression schemas (null 'PGbool)
-bool = UnsafeTypeExpression "bool"
--- | signed two-byte integer
-int2, smallint :: TypeExpression schemas (null 'PGint2)
-int2 = UnsafeTypeExpression "int2"
-smallint = UnsafeTypeExpression "smallint"
--- | signed four-byte integer
-int4, int, integer :: TypeExpression schemas (null 'PGint4)
-int4 = UnsafeTypeExpression "int4"
-int = UnsafeTypeExpression "int"
-integer = UnsafeTypeExpression "integer"
--- | signed eight-byte integer
-int8, bigint :: TypeExpression schemas (null 'PGint8)
-int8 = UnsafeTypeExpression "int8"
-bigint = UnsafeTypeExpression "bigint"
--- | arbitrary precision numeric type
-numeric :: TypeExpression schemas (null 'PGnumeric)
-numeric = UnsafeTypeExpression "numeric"
--- | single precision floating-point number (4 bytes)
-float4, real :: TypeExpression schemas (null 'PGfloat4)
-float4 = UnsafeTypeExpression "float4"
-real = UnsafeTypeExpression "real"
--- | double precision floating-point number (8 bytes)
-float8, doublePrecision :: TypeExpression schemas (null 'PGfloat8)
-float8 = UnsafeTypeExpression "float8"
-doublePrecision = UnsafeTypeExpression "double precision"
--- | currency amount
-money :: TypeExpression schema (null 'PGmoney)
-money = UnsafeTypeExpression "money"
--- | variable-length character string
-text :: TypeExpression schemas (null 'PGtext)
-text = UnsafeTypeExpression "text"
--- | fixed-length character string
-char, character
-  :: forall n schemas null. (KnownNat n, 1 <= n)
-  => TypeExpression schemas (null ('PGchar n))
-char = UnsafeTypeExpression $ "char(" <> renderNat @n <> ")"
-character = UnsafeTypeExpression $  "character(" <> renderNat @n <> ")"
--- | variable-length character string
-varchar, characterVarying
-  :: forall n schemas null. (KnownNat n, 1 <= n)
-  => TypeExpression schemas (null ('PGvarchar n))
-varchar = UnsafeTypeExpression $ "varchar(" <> renderNat @n <> ")"
-characterVarying = UnsafeTypeExpression $
-  "character varying(" <> renderNat @n <> ")"
--- | binary data ("byte array")
-bytea :: TypeExpression schemas (null 'PGbytea)
-bytea = UnsafeTypeExpression "bytea"
--- | date and time (no time zone)
-timestamp :: TypeExpression schemas (null 'PGtimestamp)
-timestamp = UnsafeTypeExpression "timestamp"
--- | date and time, including time zone
-timestampWithTimeZone :: TypeExpression schemas (null 'PGtimestamptz)
-timestampWithTimeZone = UnsafeTypeExpression "timestamp with time zone"
--- | calendar date (year, month, day)
-date :: TypeExpression schemas (null 'PGdate)
-date = UnsafeTypeExpression "date"
--- | time of day (no time zone)
-time :: TypeExpression schemas (null 'PGtime)
-time = UnsafeTypeExpression "time"
--- | time of day, including time zone
-timeWithTimeZone :: TypeExpression schemas (null 'PGtimetz)
-timeWithTimeZone = UnsafeTypeExpression "time with time zone"
--- | time span
-interval :: TypeExpression schemas (null 'PGinterval)
-interval = UnsafeTypeExpression "interval"
--- | universally unique identifier
-uuid :: TypeExpression schemas (null 'PGuuid)
-uuid = UnsafeTypeExpression "uuid"
--- | IPv4 or IPv6 host address
-inet :: TypeExpression schemas (null 'PGinet)
-inet = UnsafeTypeExpression "inet"
--- | textual JSON data
-json :: TypeExpression schemas (null 'PGjson)
-json = UnsafeTypeExpression "json"
--- | binary JSON data, decomposed
-jsonb :: TypeExpression schemas (null 'PGjsonb)
-jsonb = UnsafeTypeExpression "jsonb"
--- | variable length array
-vararray
-  :: TypeExpression schemas pg
-  -> TypeExpression schemas (null ('PGvararray pg))
-vararray ty = UnsafeTypeExpression $ renderSQL ty <> "[]"
--- | fixed length array
---
--- >>> renderSQL (fixarray @'[2] json)
--- "json[2]"
-fixarray
-  :: forall dims schemas null pg. SOP.All KnownNat dims
-  => TypeExpression schemas pg
-  -> TypeExpression schemas (null ('PGfixarray dims pg))
-fixarray ty = UnsafeTypeExpression $
-  renderSQL ty <> renderDims @dims
-  where
-    renderDims :: forall ns. SOP.All KnownNat ns => ByteString
-    renderDims =
-      ("[" <>)
-      . (<> "]")
-      . ByteString.intercalate "]["
-      . hcollapse
-      $ hcmap (SOP.Proxy @KnownNat)
-        (K . fromString . show . natVal)
-        (hpure SOP.Proxy :: NP SOP.Proxy ns)
-tsvector :: TypeExpression schemas (null 'PGtsvector)
-tsvector = UnsafeTypeExpression "tsvector"
-tsquery :: TypeExpression schemas (null 'PGtsquery)
-tsquery = UnsafeTypeExpression "tsquery"
-
--- | `pgtype` is a demoted version of a `PGType`
-class PGTyped schemas (ty :: NullityType) where
-  pgtype :: TypeExpression schemas ty
-instance PGTyped schemas (null 'PGbool) where pgtype = bool
-instance PGTyped schemas (null 'PGint2) where pgtype = int2
-instance PGTyped schemas (null 'PGint4) where pgtype = int4
-instance PGTyped schemas (null 'PGint8) where pgtype = int8
-instance PGTyped schemas (null 'PGnumeric) where pgtype = numeric
-instance PGTyped schemas (null 'PGfloat4) where pgtype = float4
-instance PGTyped schemas (null 'PGfloat8) where pgtype = float8
-instance PGTyped schemas (null 'PGmoney) where pgtype = money
-instance PGTyped schemas (null 'PGtext) where pgtype = text
-instance (KnownNat n, 1 <= n)
-  => PGTyped schemas (null ('PGchar n)) where pgtype = char @n
-instance (KnownNat n, 1 <= n)
-  => PGTyped schemas (null ('PGvarchar n)) where pgtype = varchar @n
-instance PGTyped schemas (null 'PGbytea) where pgtype = bytea
-instance PGTyped schemas (null 'PGtimestamp) where pgtype = timestamp
-instance PGTyped schemas (null 'PGtimestamptz) where pgtype = timestampWithTimeZone
-instance PGTyped schemas (null 'PGdate) where pgtype = date
-instance PGTyped schemas (null 'PGtime) where pgtype = time
-instance PGTyped schemas (null 'PGtimetz) where pgtype = timeWithTimeZone
-instance PGTyped schemas (null 'PGinterval) where pgtype = interval
-instance PGTyped schemas (null 'PGuuid) where pgtype = uuid
-instance PGTyped schemas (null 'PGjson) where pgtype = json
-instance PGTyped schemas (null 'PGjsonb) where pgtype = jsonb
-instance PGTyped schemas ty
-  => PGTyped schemas (null ('PGvararray ty)) where
-    pgtype = vararray (pgtype @schemas @ty)
-instance (SOP.All KnownNat dims, PGTyped schemas ty)
-  => PGTyped schemas (null ('PGfixarray dims ty)) where
-    pgtype = fixarray @dims (pgtype @schemas @ty)
-instance PGTyped schemas (null 'PGtsvector) where pgtype = tsvector
-instance PGTyped schemas (null 'PGtsquery) where pgtype = tsquery
-
 class Literal hask where
   literal :: hask -> Expr (null (PG hask))
 instance JSON.ToJSON hask => Literal (Json hask) where
-  literal = cast json . UnsafeExpression
+  literal = UnsafeExpression . parenthesized . (<> " :: json")
     . singleQuotedUtf8 . toStrict . JSON.encode . getJson
 instance JSON.ToJSON hask => Literal (Jsonb hask) where
-  literal = cast jsonb . UnsafeExpression
+  literal =  UnsafeExpression . parenthesized . (<> " :: jsonb")
     . singleQuotedUtf8 . toStrict . JSON.encode . getJsonb
 instance Literal Char where
   literal chr = UnsafeExpression $
