+++ squeal-postgresql/src/Squeal/PostgreSQL/Query.hs
@@ -116,6 +116,7 @@ import Squeal.PostgreSQL.Alias
 import Squeal.PostgreSQL.Expression
 import Squeal.PostgreSQL.Expression.Logic
 import Squeal.PostgreSQL.Expression.Sort
+import Squeal.PostgreSQL.Expression.Type
 import Squeal.PostgreSQL.Expression.Window
 import Squeal.PostgreSQL.List
 import Squeal.PostgreSQL.PG
